requirements.txt
Flask==3.0.3
supabase==2.4.1
beautifulsoup4==4.12.3
requests==2.31.0
python-dotenv==1.0.1
gunicorn==22.0.0
playwright==1.42.0
google-generativeai==0.8.4
flask-cors==4.0.0
